EQL: Use byte instead of int for stages
Memory usage analysis revealed that we use an `int` for the number of
stages of a sequence query, whereas we could use a `byte` instead.
With this opportunity added a check in the `Verifier` which prevents
sequence queries with more than 127 stages.

+++ x-pack/plugin/eql/src/main/java/org/elasticsearch/xpack/eql/execution/assembler/Criterion.java
@@ -17,7 +17,7 @@ import java.util.List;
 
 public class Criterion<Q extends QueryRequest> {
 
-    private final int stage;
+    private final byte stage;
     private final Q queryRequest;
     private final List<HitExtractor> keys;
     private final HitExtractor timestamp;
@@ -26,7 +26,7 @@ public class Criterion<Q extends QueryRequest> {
     private final boolean descending;
     private final int keySize;
 
-    public Criterion(int stage,
+    public Criterion(byte stage,
               Q queryRequest,
               List<HitExtractor> keys,
               HitExtractor timestamp,
@@ -47,7 +47,7 @@ public class Criterion<Q extends QueryRequest> {
         return keySize;
     }
 
-    public int stage() {
+    public byte stage() {
         return stage;
     }
